Drop system
The drop system bug has been fixed.

Bug: When there were multiple drops on the ground, the information from the first generated drop would be displayed. However, upon choosing the weapon, only the last generated drop would be sent to the Player.

Reason: The code allocated the drop at the moment when the information was generated.

Fix: The trigger that decides which drop to use has been moved to the code that detects interaction with the drop.

Still pending is the creation of a system where the drop detects when another drop is within its hitbox. This will cause both drops to move apart so that the player doesn't have trouble choosing between them.