Drop repulsion
A system has been added to make dropped items move away if they are too close. It is still necessary to calibrate the impulse force to prevent them from moving too far apart.

BUG: When drops are generated, the Circle Collider is not centered with the image.

FIX: It is necessary to change the "sprite sort point" to "pivot", but for this, the weapon must have the "icon" sprite, so it will be centered.

The only thing missing at the moment is to create a system for the player to drop the weapon when replacing it with a new one.

A possible bug is that the drops might pass through walls when moving away as they don't have a physical body (needs testing).