Drop System (Complete)
The drop system is now complete. It's possible to generate a random weapon or a specific one when the player replaces a weapon in their hand; it is automatically dropped, and it can be retrieved later. When two drops are very close to each other, they repel each other, allowing the player to choose which one to interact with.

It's possible that when repelling, drops may escape collisions, potentially going outside the map or getting stuck inside a solid object. Testing is required.